requirements.txt
# Django + Django Ninja
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-ninja>=1.0,<2.0
django-environ>=0.11,<1.0
drf-spectacular>=0.27,<1.0

# Django Packages
django-cors-headers>=4.9,<5.0
django-axes>=7.1,<8.0
django-ratelimit>=4.1,<5.0
django-simple-history>=3.1,<4.0
django-celery-results>=2.5,<3.0
django-health-check>=3.17,<4.0

# Channels + Redis
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
redis>=5.0,<6.0

# Celery
celery>=5.3,<6.0
celery-redbeat>=2.2,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# Validation & Serialization
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

# Utilities
python-dotenv>=1.0,<2.0
phonenumbers>=8.13,<9.0
 Pillow>=10.0,<11.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
faker>=23.0,<24.0
freezegun>=1.4,<2.0

# Code Quality
flake8>=7.0,<8.0
black>=24.0,<25.0
isort>=5.13,<6.0

# Misc
gunicorn>=21.0,<22.0